Big Changes to Models and DB
Made update to Person and created Corporate Model. Both now reflect the Columns inside example Excel Sheet.

If Database doesn't work, delete it and next time you run it, it will recreate it correctly.

ExcelReading File can open up file and fill all the information into a string and output it to console. Next step is to parse the data and then upload it correctly to the DB.

Updated the View Page to pull data from all the data from a Person Object