It is now uploading the data from the excel file to the DB
Now it is parsing the Data from the file and making an object to upload to the DB. Right now it is not allowing any sort of updating, so if you make changes to the Excel sheet, you need to delete the DB and run the program again.

I added the Corporate controller and view so I can display what is inside the DB.

Final steps for the Database on this project is to allow updating.